requirements.txt
torch
torchvision
Pillow
matplotlib
tk
